phase2/Claude/requirements.txt
scipy>=1.11.0
numpy>=1.24.0
Pillow>=10.0.0
pytest>=7.4.0
pytest-mock>=3.11.0
